+++ test/framework/src/main/java/org/opensearch/test/telemetry/MockTelemetry.java
@@ -15,8 +15,12 @@ import org.opensearch.telemetry.metrics.Histogram;
 import org.opensearch.telemetry.metrics.MetricsTelemetry;
 import org.opensearch.telemetry.metrics.noop.NoopCounter;
 import org.opensearch.telemetry.metrics.noop.NoopHistogram;
+import org.opensearch.telemetry.metrics.tags.Tags;
 import org.opensearch.telemetry.tracing.TracingTelemetry;
 import org.opensearch.test.telemetry.tracing.MockTracingTelemetry;
+
+import java.io.Closeable;
+import java.util.function.Supplier;
 
 /**
  * Mock {@link Telemetry} implementation for testing.
@@ -51,6 +55,11 @@ public class MockTelemetry implements Telemetry {
             @Override
             public Histogram createHistogram(String name, String description, String unit) {
                 return NoopHistogram.INSTANCE;
+            }
+
+            @Override
+            public Closeable createGauge(String name, String description, String unit, Supplier<Double> valueProvider, Tags tags) {
+                return () -> {};
             }
 
             @Override